+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## 0.9.0 (tbd)
 
-* Add `removeAllListeners` method [#12]
+* Add TypeScript support ([#13])
+* Add `removeAllListeners` method ([#12])
 
 ## 0.8.0 (Jul 10, 2018)
 
-* Rename `.onPackedMessage` into `.onUnpackedMessage` [#10]
+* Rename `.onPackedMessage` into `.onUnpackedMessage` ([#10])
 * Replace `controlled-promise` with [`promise-controller`](https://github.com/vitalets/promise-controller)
 
 ## 0.7.0 (Jun 29, 2018)
@@ -37,3 +38,4 @@ are replaced with 4 more atomic hadlers:
 [#9]: https://github.com/vitalets/websocket-as-promised/issues/9
 [#10]: https://github.com/vitalets/websocket-as-promised/issues/10
 [#12]: https://github.com/vitalets/websocket-as-promised/issues/12
+[#13]: https://github.com/vitalets/websocket-as-promised/issues/13